--- src/main/kotlin/alvarx4z/medialibrary/domain/book/Cover.kt
@@ -1,5 +1,0 @@
-package alvarx4z.medialibrary.domain.book
-
-import alvarx4z.medialibrary.domain.shared.URL
-
-data class Cover(val url: URL)
