+++ src/main/kotlin/alvarx4z/medialibrary/domain/shared/Cover.kt
@@ -1,0 +1,3 @@
+package alvarx4z.medialibrary.domain.shared
+
+data class Cover(val url: URL)
