+++ src/test/kotlin/alvarx4z/medialibrary/domain/shared/CoverTest.kt
@@ -1,14 +1,13 @@
-package alvarx4z.medialibrary.domain.book
+package alvarx4z.medialibrary.domain.shared
 
-import alvarx4z.medialibrary.BookHelper
-import alvarx4z.medialibrary.domain.shared.URL
+import alvarx4z.medialibrary.SharedHelper
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 
 internal class CoverTest {
   @Test
   fun `should instantiate a valid Cover`() {
-    val cover = BookHelper.cover()
+    val cover = SharedHelper.cover()
 
     assertThat(cover).isInstanceOf(Cover::class.java)
     assertThat(cover.url).isInstanceOf(URL::class.java)
